添加maybe-async，保留一份异步代码，同时支持同步与异步feature
