+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "aliyun-oss-rust-sdk"
-version = "0.2.1"
-edition = "2021"
+version = "0.2.2"
+edition = "2024"
 authors = ["lake"]
 description = "阿里云 © Alibaba Cloud Official Oss SDK (标准库)"
 license = "MIT"
@@ -13,21 +13,24 @@ members = ["example"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-reqwest = { version = "0.11", features = ["json"] }
+reqwest = { version = "0.12.22", optional = true, features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
 chrono = "0.4.31"
 hmac = "0.12.1"
-base64 = "0.21.5"
+base64 = "0.22.1"
 sha1 = "0.10.6"
 urlencoding = "2.1.3"
 tracing = { version = "0.1.40", optional = true }
 tracing-subscriber = { version = "0.3.18", optional = true }
 serde_json = "1.0.110"
-thiserror = "1.0.30"
+thiserror = "2.0.12"
+maybe-async = "0.2.10"
+
 [features]
-default = []
-blocking = ["reqwest/blocking"]
+default = ["reqwest/blocking"]
+async = ["reqwest"]
 debug-print = ["tracing", "tracing-subscriber"]
+
 [dev-dependencies]
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
