+++ example/Cargo.toml
@@ -7,6 +7,6 @@ publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-aliyun-oss-rust-sdk = { path = "..", features = ["debug-print"] }
+aliyun-oss-rust-sdk = { path = "..", features = ["debug-print", "async"] }
 dotenvy = "0.15"
 tokio = { version = "1.15", features = ["full"] }
