--- src/blocking/mod.rs
@@ -1,1 +1,0 @@
-mod object;
