+++ src/lib.rs
@@ -5,10 +5,6 @@ pub mod request;
 pub mod url;
 pub mod metadata;
 mod util;
-
-#[cfg(feature = "blocking")]
-pub mod blocking;
-#[cfg(not(feature = "blocking"))]
 pub mod async_impl;
 pub mod entity;
 pub mod error;
